# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API - Clientes e Lotes",
        "description": "Documentação da API (Laravel 11).",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Servidor"
        }
    ],
    "paths": {
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Listar clientes",
                "operationId": "de55d50d1b708d9e5c63ee3f8fe8f54a",
                "parameters": [
                    {
                        "name": "nome",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "maxLength": 150,
                            "minLength": 2
                        }
                    },
                    {
                        "name": "tipo_pessoa",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "FISICA",
                                "JURIDICA"
                            ]
                        }
                    },
                    {
                        "name": "cpf",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]{11}$"
                        }
                    },
                    {
                        "name": "cnpj",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "pattern": "^[0-9]{14}$"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "email"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "nome",
                                "tipo_pessoa",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "dir",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClientePage"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Filtro inválido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Criar cliente",
                "operationId": "4ac6cdfa8d6c188fa2cb4f3263d4318a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteStore"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito (CPF/CNPJ já existentes)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clientes/{id}": {
            "get": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Detalhar cliente",
                "operationId": "0d3826d79390929601421b5810b61231",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar cliente (total)",
                "operationId": "74956b09f3ea120b5dd531c73776ca11",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito (CPF/CNPJ já existentes)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Excluir cliente",
                "operationId": "d15e304c609752c9f409ab81072fb52e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Excluído",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "Clientes"
                ],
                "summary": "Atualizar cliente (parcial)",
                "operationId": "7b98a0253386e403a5a02032fceb3315",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClienteUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Cliente"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Campos mal formatados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Não encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflito (CPF/CNPJ já existentes)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Regra de validação",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Cliente": {
                "required": [
                    "id",
                    "nome",
                    "tipo_pessoa"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "endereco": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefone": {
                        "type": "string",
                        "maxLength": 30,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "tipo_pessoa": {
                        "type": "string",
                        "enum": [
                            "FISICA",
                            "JURIDICA"
                        ]
                    },
                    "cpf": {
                        "description": "11 dígitos",
                        "type": "string",
                        "nullable": true
                    },
                    "cnpj": {
                        "description": "14 dígitos",
                        "type": "string",
                        "nullable": true
                    },
                    "responsavel_nome": {
                        "type": "string",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "responsavel_cpf": {
                        "description": "11 dígitos",
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "ClienteStore": {
                "required": [
                    "nome",
                    "tipo_pessoa"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "endereco": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefone": {
                        "type": "string",
                        "maxLength": 30,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "tipo_pessoa": {
                        "type": "string",
                        "enum": [
                            "FISICA",
                            "JURIDICA"
                        ]
                    },
                    "cpf": {
                        "description": "Obrigatório se FISICA; 11 dígitos",
                        "type": "string",
                        "nullable": true
                    },
                    "cnpj": {
                        "description": "Obrigatório se JURIDICA; 14 dígitos",
                        "type": "string",
                        "nullable": true
                    },
                    "responsavel_nome": {
                        "description": "Obrigatório se JURIDICA",
                        "type": "string",
                        "nullable": true
                    },
                    "responsavel_cpf": {
                        "description": "Obrigatório se JURIDICA; 11 dígitos",
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClienteUpdate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 150
                    },
                    "endereco": {
                        "type": "string",
                        "nullable": true
                    },
                    "telefone": {
                        "type": "string",
                        "maxLength": 30,
                        "nullable": true
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "maxLength": 150,
                        "nullable": true
                    },
                    "tipo_pessoa": {
                        "type": "string",
                        "enum": [
                            "FISICA",
                            "JURIDICA"
                        ]
                    },
                    "cpf": {
                        "type": "string",
                        "nullable": true
                    },
                    "cnpj": {
                        "type": "string",
                        "nullable": true
                    },
                    "responsavel_nome": {
                        "type": "string",
                        "nullable": true
                    },
                    "responsavel_cpf": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ClientePage": {
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Cliente"
                        }
                    },
                    "links": {
                        "$ref": "#/components/schemas/PaginationLinks"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/PaginationMeta"
                    }
                },
                "type": "object"
            },
            "PaginationLinks": {
                "properties": {
                    "first": {
                        "type": "string",
                        "nullable": true
                    },
                    "last": {
                        "type": "string",
                        "nullable": true
                    },
                    "prev": {
                        "type": "string",
                        "nullable": true
                    },
                    "next": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer"
                    },
                    "from": {
                        "type": "integer",
                        "nullable": true
                    },
                    "last_page": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "per_page": {
                        "type": "integer"
                    },
                    "to": {
                        "type": "integer",
                        "nullable": true
                    },
                    "total": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ErrorResponse": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Erro de validação"
                    },
                    "errors": {
                        "type": "object",
                        "nullable": true,
                        "additionalProperties": {
                            "schema": "CommonSchemas"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}